refactor: rename StepLine to StepMarkerConnector and move to StepMarker component
