--- components/ProcessSteps/src/StepLine.scss
@@ -1,71 +1,0 @@
-.denhaag-process-steps__step-line {
-  position: absolute;
-  left: calc(var(--denhaag-step-marker-size) / 2);
-  top: 0;
-  height: 100%;
-  outline: var(--denhaag-process-steps-step-line-width, 1px) var(--denhaag-process-steps-step-line-style, dashed)
-    var(--denhaag-process-steps-step-line-color);
-}
-
-.denhaag-process-steps__step-line--main-to-main {
-  padding-block-end: var(--denhaag-process-steps-step-distance);
-}
-
-.denhaag-process-steps__step-line--nested-to-nested {
-  top: var(--denhaag-process-steps-sub-step-marker-size);
-  padding-block-end: var(--denhaag-process-steps-sub-step-distance);
-}
-
-.denhaag-process-steps__step-line--nested-to-main {
-  top: var(--denhaag-process-steps-sub-step-marker-size);
-  padding-block-end: var(--denhaag-process-steps-step-distance);
-}
-
-.denhaag-process-steps__step-line--not-checked {
-  outline-width: var(
-    --denhaag-process-steps-step-line-not-checked-outline-width,
-    var(--denhaag-process-steps-step-line-outline-width, 1px)
-  );
-  outline-style: var(--denhaag-process-steps-step-line-not-checked-outline-style, dashed);
-  outline-color: var(
-    --denhaag-process-steps-step-line-not-checked-outline-color,
-    var(--denhaag-process-steps-step-line-outline-color)
-  );
-}
-
-.denhaag-process-steps__step-line--checked,
-.denhaag-process-steps__step-line--current {
-  outline-width: var(
-    --denhaag-process-steps-step-line-checked-outline-width,
-    var(--denhaag-process-steps-step-line-outline-width, 1px)
-  );
-  outline-style: var(--denhaag-process-steps-step-line-checked-outline-style, solid);
-  outline-color: var(
-    --denhaag-process-steps-step-line-checked-outline-color,
-    var(--denhaag-process-steps-step-line-outline-color)
-  );
-}
-
-.denhaag-process-steps__step-line--warning {
-  outline-width: var(
-    --denhaag-process-steps-step-line-warning-outline-width,
-    var(--denhaag-process-steps-step-line-outline-width, 1px)
-  );
-  outline-style: var(--denhaag-process-steps-step-line-warning-outline-style, solid);
-  outline-color: var(
-    --denhaag-process-steps-step-line-warning-outline-color,
-    var(--denhaag-process-steps-step-line-outline-color)
-  );
-}
-
-.denhaag-process-steps__step-line--error {
-  outline-width: var(
-    --denhaag-process-steps-step-line-error-outline-width,
-    var(--denhaag-process-steps-step-line-outline-width, 1px)
-  );
-  outline-style: var(--denhaag-process-steps-step-line-error-outline-style, solid);
-  outline-color: var(
-    --denhaag-process-steps-step-line-error-outline-color,
-    var(--denhaag-process-steps-step-line-outline-color)
-  );
-}
