+++ components/ProcessSteps/src/index.scss
@@ -1,6 +1,5 @@
 @import "./Step";
 @import "./StepList";
-@import "./StepLine";
 @import "./StepHeader";
 @import "./StepHeading";
 @import "./StepDetails";
