+++ components/StepMarker/src/StepMarker.scss
@@ -1,0 +1,149 @@
+.denhaag-step-marker {
+  position: relative;
+  z-index: 1;
+  align-items: center;
+  background-color: var(--denhaag-step-marker-background-color);
+  block-size: var(--denhaag-step-marker-size);
+  border-style: solid;
+  border-width: var(--denhaag-step-marker-border-width);
+  border-color: var(--denhaag-step-marker-border-color, transparent);
+  border-radius: var(--denhaag-step-marker-border-radius, 50%);
+  box-sizing: border-box;
+  color: var(--denhaag-step-marker-color);
+  display: flex;
+  flex-grow: 0;
+  font-family: var(--denhaag-step-marker-font-family);
+  font-size: var(--denhaag-step-marker-font-size);
+  font-weight: var(--denhaag-step-marker-font-weight);
+  inline-size: var(--denhaag-step-marker-size);
+  justify-content: center;
+  margin-block-end: var(--denhaag-step-marker-margin);
+  margin-block-start: var(--denhaag-step-marker-margin);
+  padding-inline-end: 0;
+  min-inline-size: var(--denhaag-step-marker-min-size, var(--denhaag-step-marker-size));
+  min-block-size: var(--denhaag-step-marker-min-size, var(--denhaag-step-marker-size));
+
+  .denhaag-icon {
+    block-size: var(--denhaag-step-marker-icon-size);
+    inline-size: var(--denhaag-step-marker-icon-size);
+  }
+}
+
+.denhaag-step-marker--nested {
+  block-size: var(--denhaag-step-marker-nested-size);
+  min-block-size: var(--denhaag-step-marker-nested-min-size, var(--denhaag-step-marker-nested-size));
+  min-inline-size: var(--denhaag-step-marker-nested-min-size, var(--denhaag-step-marker-nested-size));
+  inline-size: var(--denhaag-step-marker-nested-size);
+  border-width: var(--denhaag-step-marker-nested-border-width, var(--denhaag-step-marker-border-width));
+  margin-inline-start: calc((var(--denhaag-step-marker-size) - var(--denhaag-step-marker-nested-size)) / 2);
+
+  .denhaag-icon {
+    block-size: var(--denhaag-step-marker-nested-icon-size);
+    inline-size: var(--denhaag-step-marker-nested-icon-size);
+  }
+}
+
+.denhaag-step-marker--not-checked {
+  background-color: var(--denhaag-step-marker-not-checked-background-color);
+  border-color: var(--denhaag-step-marker-not-checked-border-color);
+  color: var(--denhaag-step-marker-not-checked-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-not-checked-nested-background-color,
+      var(--denhaag-step-marker-not-checked-background-color)
+    );
+    border-color: var(
+      --denhaag-step-marker-not-checked-nested-border-color,
+      var(--denhaag-step-marker-not-checked-border-color)
+    );
+    color: var(--denhaag-step-marker-not-checked-nested-color, var(--denhaag-step-marker-not-checked-color));
+  }
+}
+
+.denhaag-step-marker--checked {
+  background-color: var(--denhaag-step-marker-checked-background-color);
+  border-color: var(--denhaag-step-marker-checked-border-color);
+  color: var(--denhaag-step-marker-checked-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-checked-nested-background-color,
+      var(--denhaag-step-marker-checked-background-color)
+    );
+    border-color: var(
+      --denhaag-step-marker-checked-nested-border-color,
+      var(--denhaag-step-marker-checked-border-color)
+    );
+    color: var(--denhaag-step-marker-checked-nested-color, var(--denhaag-step-marker-checked-color));
+  }
+}
+
+.denhaag-step-marker--current {
+  background-color: var(--denhaag-step-marker-current-background-color);
+  border-color: var(--denhaag-step-marker-current-border-color);
+  color: var(--denhaag-step-marker-current-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-current-nested-background-color,
+      var(--denhaag-step-marker-current-background-color)
+    );
+    border-color: var(
+      --denhaag-step-marker-current-nested-border-color,
+      var(--denhaag-step-marker-current-border-color)
+    );
+    color: var(--denhaag-step-marker-current-nested-color, var(--denhaag-step-marker-current-color));
+  }
+}
+
+.denhaag-step-marker--disabled {
+  background-color: var(--denhaag-step-marker-disabled-background-color);
+  border-color: var(--denhaag-step-marker-disabled-border-color);
+  color: var(--denhaag-step-marker-disabled-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-disabled-nested-background-color,
+      var(--denhaag-step-marker-disabled-background-color)
+    );
+    border-color: var(
+      --denhaag-step-marker-disabled-nested-border-color,
+      var(--denhaag-step-marker-disabled-border-color)
+    );
+    color: var(--denhaag-step-marker-disabled-nested-color, var(--denhaag-step-marker-disabled-color));
+  }
+}
+
+.denhaag-step-marker--warning {
+  background-color: var(--denhaag-step-marker-warning-background-color);
+  border-color: var(--denhaag-step-marker-warning-border-color);
+  color: var(--denhaag-step-marker-warning-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-warning-nested-background-color,
+      var(--denhaag-step-marker-warning-background-color)
+    );
+    border-color: var(
+      --denhaag-step-marker-warning-nested-border-color,
+      var(--denhaag-step-marker-warning-border-color)
+    );
+    color: var(--denhaag-step-marker-warning-nested-color, var(--denhaag-step-marker-warning-color));
+  }
+}
+
+.denhaag-step-marker--error {
+  background-color: var(--denhaag-step-marker-error-background-color);
+  border-color: var(--denhaag-step-marker-error-border-color);
+  color: var(--denhaag-step-marker-error-color);
+
+  &.denhaag-step-marker--nested {
+    background-color: var(
+      --denhaag-step-marker-error-nested-background-color,
+      var(--denhaag-step-marker-error-background-color)
+    );
+    border-color: var(--denhaag-step-marker-error-nested-border-color, var(--denhaag-step-marker-error-border-color));
+    color: var(--denhaag-step-marker-error-nested-color, var(--denhaag-step-marker-error-color));
+  }
+}
