+++ components/StepMarker/src/StepMarkerConnector.tsx
@@ -1,0 +1,34 @@
+import React, { HTMLAttributes } from 'react';
+
+import clsx from 'clsx';
+
+type MarkerType = 'main' | 'nested';
+type Appearance = 'checked' | 'not-checked' | 'current' | 'warning' | 'error';
+
+interface StepMarkerConnectorProps extends Omit<HTMLAttributes<HTMLDivElement>, 'children'> {
+  from: MarkerType;
+  to: MarkerType;
+  appearance?: Appearance;
+}
+
+export const StepMarkerConnector: React.FC<StepMarkerConnectorProps> = ({
+  className,
+  appearance = 'not-checked',
+  from,
+  to,
+  ...props
+}: StepMarkerConnectorProps) => {
+  return (
+    <div
+      {...props}
+      className={clsx(
+        'denhaag-step-marker__connector',
+        `denhaag-step-marker__connector--${appearance}`,
+        `denhaag-step-marker__connector--${from}-to-${to}`,
+        className,
+      )}
+    ></div>
+  );
+};
+
+export default StepMarkerConnector;
